Update paths, package names and resource bundles to match the new samples structure
