+++ .gitignore
@@ -6,3 +6,4 @@
 **/.project
 **/.settings
 **/.vscode
+**/.classpath
