+++ portal/7.1/java8/code/modules/applications/portlets/generic-ds-plain/.sample
@@ -6,7 +6,7 @@ displayFiles:
     path: README.md
   - name: Portlet class
     highlighting: java 
-    path: src/main/java/com/liferay/samples/portlet/ds/DSPortlet.java
+    path: src/main/java/com/liferay/code/samples/portal/modules/applications/portlets/generic_ds_plain/DSPortlet.java
   - name: Bnd configuration
     highlighting: properies 
     path: bnd.bnd
