+++ portal/7.1/java8/code/modules/applications/portlets/generic-ds-plain/src/main/java/com/liferay/code/samples/portal/modules/applications/portlets/generic_ds_plain/DSPortlet.java
@@ -14,11 +14,16 @@
  * limitations under the License.
  */
 
-package com.liferay.blade.samples.portlet.jsp;
+package com.liferay.code.samples.portal.modules.applications.portlets.generic_ds_plain;
 
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
+import java.io.IOException;
+import java.io.PrintWriter;
 
+import javax.portlet.GenericPortlet;
 import javax.portlet.Portlet;
+import javax.portlet.PortletException;
+import javax.portlet.RenderRequest;
+import javax.portlet.RenderResponse;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -28,17 +33,22 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	immediate = true,
 	property = {
-		"com.liferay.portlet.css-class-wrapper=portlet-jsp",
 		"com.liferay.portlet.display-category=category.sample",
-		"com.liferay.portlet.header-portlet-css=/css/main.css",
 		"com.liferay.portlet.instanceable=true",
-		"javax.portlet.display-name=Blade JSP Portlet",
-		"javax.portlet.init-param.template-path=/",
-		"javax.portlet.init-param.view-template=/view.jsp",
-		"javax.portlet.resource-bundle=content.Language",
+		"javax.portlet.display-name=DS Portlet",
 		"javax.portlet.security-role-ref=power-user,user"
 	},
 	service = Portlet.class
 )
-public class JSPPortlet extends MVCPortlet {
+public class DSPortlet extends GenericPortlet {
+
+	@Override
+	protected void doView(RenderRequest request, RenderResponse response)
+		throws IOException, PortletException {
+
+		PrintWriter printWriter = response.getWriter();
+
+		printWriter.print("DS Portlet - Hello World!");
+	}
+
 }
