+++ portal/7.1/java8/code/modules/applications/portlets/mvc-ds-jsp/src/main/resources/content/Language.properties
@@ -1,5 +1,6 @@
-javax.portlet.title.blade_portlet_jsp_JSPPortlet=BLADE JSP Portlet
-javax.portlet.display-name.blade_portlet_jsp_JSPPortlet=BLADE JSP
-javax.portlet.short-title.blade_portlet_jsp_JSPPortlet=BLADE JSP
-javax.portlet.keywords.blade_portlet_jsp_JSPPortlet=BLADE,jsp
-blade_portlet_jsp_JSPPortlet.caption=Hello from BLADE JSP!
+javax.portlet.title.mvc_ds_jsp_portlet=Liferay MVC Portlet using DS and JSP
+javax.portlet.display-name.mvc_ds_jsp_portlet=Liferay MVC Portlet using DS and JSP
+javax.portlet.short-title.mvc_ds_jsp_portlet=Liferay MVC Portlet using DS and JSP
+javax.portlet.keywords.mvc_ds_jsp_portlet=MVC, DS, JSP
+
+mvc_ds_jsp_portlet.caption=Hello from Liferay MVC Portlet using DS and JSP!
