+++ portal/7.2/java8/code/modules/applications/portlets/mvc-ds-jsp/src/main/java/com/liferay/code/samples/portal/modules/applications/portlets/mvc_ds_jsp/JSPPortlet.java
@@ -14,16 +14,11 @@
  * limitations under the License.
  */
 
-package com.liferay.samples.portlet.ds;
+package com.liferay.code.samples.portal.modules.applications.portlets.mvc_ds_jsp;
 
-import java.io.IOException;
-import java.io.PrintWriter;
+import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
 
-import javax.portlet.GenericPortlet;
 import javax.portlet.Portlet;
-import javax.portlet.PortletException;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -33,22 +28,20 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	immediate = true,
 	property = {
+		"com.liferay.portlet.css-class-wrapper=portlet-jsp",
 		"com.liferay.portlet.display-category=category.sample",
+		"com.liferay.portlet.header-portlet-css=/css/main.css",
 		"com.liferay.portlet.instanceable=true",
-		"javax.portlet.display-name=DS Portlet",
+		"javax.portlet.name=" + JSPPortlet.NAME,
+		"javax.portlet.init-param.template-path=/",
+		"javax.portlet.init-param.view-template=/view.jsp",
+		"javax.portlet.resource-bundle=content.Language",
 		"javax.portlet.security-role-ref=power-user,user"
 	},
 	service = Portlet.class
 )
-public class DSPortlet extends GenericPortlet {
+public class JSPPortlet extends MVCPortlet {
 
-	@Override
-	protected void doView(RenderRequest request, RenderResponse response)
-		throws IOException, PortletException {
-
-		PrintWriter printWriter = response.getWriter();
-
-		printWriter.print("DS Portlet - Hello World!");
-	}
+	public static final String NAME = "mvc_ds_jsp_portlet";
 
 }
