+++ portal/7.2/java8/code/modules/applications/portlets/mvc-ds-jsp/src/main/resources/META-INF/resources/view.jsp
@@ -19,5 +19,5 @@
 <%@ include file="/init.jsp" %>
 
 <p>
-	<b class="redBackground"><liferay-ui:message key="blade_portlet_jsp_JSPPortlet.caption" /></b>
+	<b class="redBackground"><liferay-ui:message key="mvc_ds_jsp_portlet.caption" /></b>
 </p>
